// web/src/shared/api/generated/model/entitlementResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { PlanTier } from './planTier';
import type { SubscriptionResult } from './subscriptionResult';

export interface EntitlementResult {
  tier: PlanTier;
  hasAskTeacher: boolean;
  canTakeExams: boolean;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  dailyQuizQuestionLimit: number | string | null;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  dailyAvatarMessageLimit: number | string;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  openLessonsPerUnit: number | string | null;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  monthlyAskTeacherQuestionLimit: number | string;
  subscriptions: SubscriptionResult[];
}