// web/src/shared/api/generated/model/planCatalogueResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { AskTeacherPlanResult } from './askTeacherPlanResult';
import type { BasePlanResult } from './basePlanResult';
import type { FreePlanResult } from './freePlanResult';

export interface PlanCatalogueResult {
  free: FreePlanResult;
  base: BasePlanResult;
  askTeacher: AskTeacherPlanResult;
}