// web/src/shared/api/generated/model/subscriptionStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export type SubscriptionStatus = typeof SubscriptionStatus[keyof typeof SubscriptionStatus];


export const SubscriptionStatus = {
  Active: 'Active',
  PastDue: 'PastDue',
  Cancelled: 'Cancelled',
  Expired: 'Expired',
} as const;